Capstone-master/requirements.txt
Flask
gunicorn
numpy
pandas
tensorflow
pillow